Refactor blog components and restructure imports
