+++ app/blog/blog-card.jsx
