+++ app/education/index.jsx
@@ -1,15 +1,15 @@
 // @flow strict
-
-import { experiences } from "@/utils/data/experience";
 import Image from "next/image";
-import { BsPersonWorkspace } from "react-icons/bs";
-import AnimationLottie from "../../helper/animation-lottie";
-import GlowCard from "../../helper/glow-card";
-import experience from '/public/lottie/code.json';
 
-function Experience() {
+import { educations } from "@/utils/data/educations";
+import { BsPersonWorkspace } from "react-icons/bs";
+import AnimationLottie from "../components/helper/animation-lottie";
+import lottieFile from '/public/lottie/study.json';
+import GlowCard from "../components/helper/glow-card";
+
+function Education() {
   return (
-    <div id="experience" className="relative z-50 border-t my-12 lg:my-24 border-[#25213b]">
+    <div id="education" className="relative z-50 border-t my-12 lg:my-24 border-[#25213b]">
       <Image
         src="/section.svg"
         alt="Hero"
@@ -17,12 +17,17 @@ function Experience() {
         height={795}
         className="absolute top-0 -z-10"
       />
+      <div className="flex justify-center -translate-y-[1px]">
+        <div className="w-3/4">
+          <div className="h-[1px] bg-gradient-to-r from-transparent via-violet-500 to-transparent  w-full" />
+        </div>
+      </div>
 
       <div className="flex justify-center my-5 lg:py-8">
         <div className="flex  items-center">
           <span className="w-24 h-[2px] bg-[#1a1443]"></span>
           <span className="bg-[#1a1443] w-fit text-white p-2 px-5 text-xl rounded-md">
-            Experiences
+            Educations
           </span>
           <span className="w-24 h-[2px] bg-[#1a1443]"></span>
         </div>
@@ -31,17 +36,17 @@ function Experience() {
       <div className="py-8">
         <div className="grid grid-cols-1 lg:grid-cols-2 gap-8 lg:gap-16">
           <div className="flex justify-center items-start">
-            <div className="w-full h-full">
-              <AnimationLottie animationPath={experience} />
+            <div className="w-3/4 h-3/4">
+              <AnimationLottie animationPath={lottieFile} />
             </div>
           </div>
 
           <div>
             <div className="flex flex-col gap-6">
               {
-                experiences.map(experience => (
-                  <GlowCard key={experience.id} identifier={`experience-${experience.id}`}>
-                    <div className="p-3 relative">
+                educations.map(education => (
+                  <GlowCard key={education.id} identifier={`education-${education.id}`}>
+                    <div className="p-3 relative text-white">
                       <Image
                         src="/blur-23.svg"
                         alt="Hero"
@@ -51,7 +56,7 @@ function Experience() {
                       />
                       <div className="flex justify-center">
                         <p className="text-xs sm:text-sm text-[#16f2b3]">
-                          {experience.duration}
+                          {education.duration}
                         </p>
                       </div>
                       <div className="flex items-center gap-x-8 px-3 py-5">
@@ -60,11 +65,9 @@ function Experience() {
                         </div>
                         <div>
                           <p className="text-base sm:text-xl mb-2 font-medium uppercase">
-                            {experience.title}
+                            {education.title}
                           </p>
-                          <p className="text-sm sm:text-base">
-                            {experience.company}
-                          </p>
+                          <p className="text-sm sm:text-base">{education.institution}</p>
                         </div>
                       </div>
                     </div>
@@ -79,4 +82,4 @@ function Experience() {
   );
 };
 
-export default Experience;
+export default Education;
