+++ app/hero-section/index.jsx
