+++ app/page.js
@@ -1,12 +1,12 @@
 import { personalData } from "@/utils/data/personal-data";
-import AboutSection from "./components/homepage/about";
-import Blog from "./components/homepage/blog";
-import ContactSection from "./components/homepage/contact";
-import Education from "./components/homepage/education";
-import Experience from "./components/homepage/experience";
-import HeroSection from "./components/homepage/hero-section";
-import Projects from "./components/homepage/projects";
-import Skills from "./components/homepage/skills";
+import AboutSection from "./about";
+import Blog from "./blog";
+import ContactSection from "./contact";
+import Education from "./education";
+import Experience from "./experience";
+import HeroSection from "./hero-section";
+import Projects from "./projects";
+import Skills from "./skills";
 
 async function getData() {
   const res = await fetch(`https://dev.to/api/articles?username=${personalData.devUsername}`)
