+++ app/projects/project-card.jsx
